requirements.txt
django>=4.2, <4.3
djangorestframework>=3.14, <3.15
black==23.7.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
ruff==0.0.285